Added dynamic arrays to shadervar
Former-commit-id: ec57f7c0d98ea98bd8324f08873200aebf0d8e9c
Former-commit-id: 1cf0f42ffa9eb27600eb5bee02c2e54ce05d8818 [formerly b521771f9b7363e494a9e0294f43dfb57c551d96]
Former-commit-id: 48c63d2cfcf767c798f5c1771fef8b2dbe3adc61

+++ modules/renderer/source/hiprenderer/vertex.d
@@ -290,7 +290,7 @@ class HipVertexArrayObject
             obj.appendAttribute(3, FLOAT, float.sizeof, "vPosition") //X, Y, Z
                .appendAttribute(4, FLOAT, float.sizeof, "vColor") //R, G, B, A
                .appendAttribute(2, FLOAT, float.sizeof, "vTexST") //S, T (Texture coordinates)
-               .appendAttribute(1, INT  , int.sizeof  , "vTexID");
+               .appendAttribute(1, FLOAT, float.sizeof, "vTexID");
         }
         return obj;
     }
